Support CAPI v1.11.1 v1beta2 resources (#313)
Cluster API resources change in the move to the v1beta2 api.

Notably, infrastructureRef no longer contains the apiVersion field,
which is replaced with apiGroup, and the API version removed from the
end of the string, which means that it is no longer immediately
available to fetch_ref().

This PR takes the API group and uses easykube api_preferred_version()
to determine the preferred api version of the apiGroup, if apiGroup is
available on the ref that we are working on.

If apiGroup is not available, the original behaviour (which is to take
apiVersion, and default to v1 to match Secret if none is supplied)
is retained, which should make this change compatible with
cluster.x-k8s.io/v1beta1 and cluster.x-k8s.io/v1beta2 resources.